+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **mappings:** moved q to <leader>q
 - **snippet:** added prequire snippet lua
 - **snippet:** more snippet for go
+- **snippets:** added dynamic preq snippet
 - **snippets:** added more lua snippets